+++ src/v1/Archetypes/FunctionTypes/TypeGuarantee/TypeGuarantee.ts
@@ -37,7 +37,7 @@ import { TypeGuaranteeOptions } from "./TypeGuaranteeOptions";
  * A `TypeGuarantee` inspects a piece of data to see if the data is the given
  * type.
  *
- * If the given data is the given type, the function returns `true`.
+ * If the given data is the given type, the function returns the given data.
  * It passes an error to the user-supplied `onError()` handler otherwise.
  * The `onError()` handler will throw an exception.
  *
@@ -57,5 +57,5 @@ import { TypeGuaranteeOptions } from "./TypeGuaranteeOptions";
  * @param options.path
  * We use this in thrown Errors to provide a meaningful name for `input`
  */
-export type TypeGuarantee<T, OPT extends object = object>
-    = (input: unknown, options?: Partial<TypeGuaranteeOptions> & Partial<OPT>) => T;
+export type TypeGuarantee<T, OPT extends TypeGuaranteeOptions = TypeGuaranteeOptions>
+    = (input: unknown, options?: Partial<OPT>) => T;